Update fble to work on aarch64.
Switching from a 32 bit OS to a 64 bit OS for my Raspberry Pi.

  fble bench interpreted: 6m39s ==> 7m06s
  fble bench compiled:    5m03s ==> 5m58s

+++ fble/src/typecheck.c
@@ -335,12 +335,12 @@ static void ReportError(FbleLoc loc, const char* fmt, ...)
   va_start(ap, fmt);
 
   for (const char* p = strchr(fmt, '%'); p != NULL; p = strchr(fmt, '%')) {
-    fprintf(stderr, "%.*s", p - fmt, fmt);
+    fprintf(stderr, "%.*s", (int)(p - fmt), fmt);
 
     switch (*(p + 1)) {
       case 'i': {
         size_t x = va_arg(ap, size_t);
-        fprintf(stderr, "%d", x);
+        fprintf(stderr, "%zd", x);
         break;
       }
